+++ src/main/java/security/Checks.kt
@@ -19,6 +19,8 @@ object Checks {
     val MakoTemplateFilterCheck = CheckType("MK100", "Mako does not inspect or sanitize input by default, leaving rendered templates open to XSS. Use default_filters=['h'].")
     val SqlInjectionCheck = CheckType("SQL100", "Possible SQL injection within String format")
     val AssertCheck = CheckType("AST100", "Asserts should only be used in tests. Asserts are typically bypassed in a production environment.")
+    val TryExceptPassCheck = CheckType("TRY100", "Ignoring exceptions without either logging or handling is not considered good security practice.")
+    val TryExceptContinueCheck = CheckType("TRY101", "Ignoring exceptions without either logging or handling is not considered good security practice.")
 
     class CheckType(var Code: String, var Message: String) {
         override fun toString(): String {