+++ src/test/java/security/validators/AssertInspectionTest.kt
@@ -1,24 +1,9 @@
 package security.validators
 
-import com.intellij.codeInspection.LocalInspectionToolSession
-import com.intellij.codeInspection.LocalQuickFix
-import com.intellij.codeInspection.ProblemHighlightType
-import com.intellij.codeInspection.ProblemsHolder
-import com.intellij.openapi.application.ApplicationManager
-import com.intellij.psi.PsiElement
-import com.intellij.psi.util.PsiTreeUtil
-import com.jetbrains.python.PythonFileType
-import com.jetbrains.python.inspections.PyInspection
-import com.jetbrains.python.inspections.PyInspectionVisitor
-import com.jetbrains.python.psi.PyAssertStatement
-import com.jetbrains.python.psi.PyFormattedStringElement
-import com.nhaarman.mockitokotlin2.*
-import org.jetbrains.annotations.NotNull
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
-import org.mockito.Mockito
 import security.Checks
 import security.SecurityTestTask
 
@@ -44,7 +29,7 @@ class AssertInspectionTest: SecurityTestTask() {
         var code = """
             assert 1 == 1
         """.trimIndent()
-        testAssert(code, 0, Checks.AssertCheck, "test_foo.py", AssertInspection())
+        testAssertStatement(code, 0, Checks.AssertCheck, "test_foo.py", AssertInspection())
     }
 
     @Test
@@ -52,28 +37,6 @@ class AssertInspectionTest: SecurityTestTask() {
         var code = """
             assert 1 == 1
         """.trimIndent()
-        testAssert(code, 1, Checks.AssertCheck, "my_file.py", AssertInspection())
-    }
-
-    fun <inspector: PyInspection>testAssert(code: String, times: Int = 1, check: Checks.CheckType, filename: String = "test.py", instance: inspector){
-        ApplicationManager.getApplication().runReadAction {
-            val mockHolder = mock<ProblemsHolder> {
-                on { registerProblem(any<PsiElement>(), eq(check.getDescription()), eq(ProblemHighlightType.WEAK_WARNING)) } doAnswer {}
-            }
-            val testFile = this.createLightFile(filename, PythonFileType.INSTANCE.language, code);
-            val mockLocalSession = mock<LocalInspectionToolSession> {
-                on { file } doReturn (testFile)
-            }
-            assertNotNull(testFile)
-            val testVisitor = instance.buildVisitor(mockHolder, true, mockLocalSession) as PyInspectionVisitor
-
-            val expr: @NotNull MutableCollection<PyAssertStatement> = PsiTreeUtil.findChildrenOfType(testFile, PyAssertStatement::class.java)
-            assertNotNull(expr)
-            expr.forEach { e ->
-                testVisitor.visitPyAssertStatement(e)
-            }
-            Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), eq(ProblemHighlightType.WEAK_WARNING))
-            Mockito.verify(mockLocalSession, Mockito.times(1)).file
-        }
+        testAssertStatement(code, 1, Checks.AssertCheck, "my_file.py", AssertInspection())
     }
 }